syscall refactor! Syscalls are now defined in meta/syscalls.toml
Also, added an automatically generated sysbind.h that automatically
has all syscalls defined

+++ kernel/syscall/user.cpp
@@ -1,16 +1,16 @@
 #include <syscall.h>
 #include <cpu.h>
 
-long sys::getuid(void) {
+int sys::getuid(void) {
   return curproc->user.uid;
 }
-long sys::geteuid(void) {
+int sys::geteuid(void) {
   return curproc->user.euid;
 }
 
-long sys::getgid(void) {
+int sys::getgid(void) {
   return curproc->user.gid;
 }
-long sys::getegid(void) {
+int sys::getegid(void) {
   return curproc->user.egid;
 }
